refactor: move the controller-runtime helper out of the shared util package
util/controllerruntime.go held a single function, MustSetupWithManager, with a
single caller in manager/start.go. Because it lived in the shared util package
it pulled controller-runtime into the import graph of every package that used
any util helper -- including config, and through config the topology compiler.

Give it its own leaf package. util is now free of controller-runtime, which
takes 44 controller-runtime and apiserver packages out of the transitive
dependencies of config and of anything that imports it for compilation rather
than for reconciliation.
